Sort alternates, to extract more common prefixes
Example: `abc|x|abd` - sorting allows the `ab`
prefix to be extracted, so regexp matching is more
efficient.

Where one alternate is already a prefix of another,
avoid moving the shorter one earlier in the list
(where it would be preferred), by using stable
sort and comparing up to the shorter length.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>